refactor(tests): TC-07 — extrae ChileHubDataManager y GeoCache a test_data_manager.py
test_chile_hub.py (2532 lineas) exigia make build previo para todo. Las
clases ChileHubDataManagerUnitTests y GeoCacheIntegrityTests no leen el
directorio de artefactos normalizados (temp dirs, Mocks, fixtures
sinteticos) — pasan a tests/test_data_manager.py (445 lineas, 19 tests),
que corre sin build. Incluye los tests de hardening A2 que viven en
test_chile_hub.py (cache clear path hermano + validate_user_data via
self.root_dir) y el ajuste del import de _main (chile_hub.cli) por el
refactor B1.

AGENTS.md §8: entrada nueva para test_data_manager.py en la tabla
autogenerada (doc_sync) y 13 archivos en el arbol.
